tidies server.js and puts handlers in their own directory
